+++ CHANGELOG.rst
@@ -29,6 +29,7 @@ Added
 * Command to export `Organisations` and `Persons` to CSV files.
 * Command to export `Events` to a CSV file.
 * Docs: data model fields concordance.
+* Command to export `Places` to a CSV file.
 
 Changed
 ~~~~~~~